pyproject.toml
[project]
name = "oneqlick-backend"
version = "0.1.0"
description = "OneQlick Food Delivery Backend API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "psycopg[binary]>=3.2.5",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.1",
    "pytz>=2023.3",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
    "httpx>=0.28.1",
    "redis>=5.0.0",
    "bcrypt>=4.0.0"
]
